matchsite/requirement.txt
certifi==2018.8.24
Django==2.1.3
django-crispy-forms==1.7.2
django-filter==2.0.0
django-simple-cookie-consent==0.1.1
djangorestframework==3.9.0
gunicorn==19.9.0
Markdown==3.0.1
Pillow==5.3.0
Python==3.5.2
pytz==2018.7
whitenoise==4.1.2
Conda==4.3.16